feat: added proper subcommand setup for debug stuff
